Funciones utils + Query con poll id
